requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
nagisa==0.2.10
pykakasi==2.2.1
jamdict==0.1a11
requests==2.31.0
langgraph==0.0.26
langchain-core==0.1.10
python-dotenv==1.0.0